fix migration but ref doesn't work

+++ database/migrations/2020_03_31_210306_create_replies_table.php
@@ -17,10 +17,10 @@ class CreateRepliesTable extends Migration
             $table->id();
             $table->text('body');
 
-            $table->integer('question_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('question_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
 
-            $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade'); 
+            $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
             $table->timestamps();
         });
     }
