+++ database/migrations/2020_03_31_210353_create_likes_table.php
@@ -15,8 +15,8 @@ class CreateLikesTable extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->id();
-            $table->integer('reply_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('reply_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->timestamps();
         });
     }
